+++ src/routes/Authorize/index.ts
@@ -4,3 +4,4 @@ export { default as AuthorizeRedirect } from './AuthorizeRedirect';
 export { IntroRoute } from './IntroRoute';
 export { LoginRoute } from './LoginRoute';
 export { OtpRoute } from './OtpRoute';
+export { SocialAuthRoute } from './SocialAuthRoute';